MySql/requirements.txt
mysql-connector-python==8.0.15
protobuf==3.7.1
six==1.12.0
